requirements.txt
transformers>=4.30.0
torch>=2.0.0
datasets>=2.12.0
pyyaml>=6.0
streamlit>=1.22.0
pandas>=1.5.0
numpy>=1.24.0 